% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{app_server}
\alias{app_server}
\title{Application server-side}
\usage{
app_server(input, output, session)
}
\arguments{
\item{input, output, session}{Internal parameters for {shiny}.}
}
\description{
Application server-side
}
