% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello_srv}
\alias{hello_srv}
\title{Server code for our hello world shiny app}
\usage{
hello_srv(input, output, session)
}
\arguments{
\item{input, output, session}{Internal parameters for {shiny}.}
}
\description{
This contains boiler plate code from https://shiny.rstudio.com/gallery/#demos
}
